Rename io.bitsquare.{di=>app}.BitsquareModule
