+++ src/test/java/io/bitsquare/gui/ViewLoaderTest.java
@@ -18,7 +18,7 @@
 package io.bitsquare.gui;
 
 import io.bitsquare.FatalException;
-import io.bitsquare.di.BitsquareModule;
+import io.bitsquare.app.BitsquareModule;
 
 import com.google.inject.Guice;
 import com.google.inject.Injector;
